Add clang-format/clang-tidy tooling and fix the clang-tidy findings
Bring over the style setup of ivan-pi/gpu-assembly: pre-commit drives
clang-format (pinned) and the whitespace checks, clang-tidy is a manual
stage against a compile database, a Style workflow runs both in CI, and
Claude Code gets a PostToolUse hook that formats every edit plus a
session-start hook that provisions the tools.

The existing .clang-format (C++ Core Guidelines NL.17) stays; the tree was
already clean under it. .clang-tidy carries the same check set as
gpu-assembly, with HeaderFilterRegex widened to include/.

clang-tidy runs with the system clang-tidy against a clang-configured
tree (build-tidy) rather than the pinned wheel: it parses with clang's
front end, which cannot read g++'s omp.h, and a clang-tidy of another
release does not find this release's omp.h either. libomp-dev supplies
the header; CI and the session hook install it.

Findings fixed: nested namespaces concatenated, is_floating_point_v,
explicit uint8_t bases on the three selector enums, a redundant inline,
a range-for and an empty() in the benchmarks, info printed through long
(MKL_INT is int under LP64, so the (int) cast was redundant there and
would narrow under ILP64), and NOLINT fences on the test dispatch macros,
whose T argument is a type name and cannot be parenthesized.

+++ tests/test_cqr_potrf_mkl.cpp
@@ -80,7 +80,7 @@ int suite1(MKL_LAYOUT layout, MKL_UPLO uplo, int nm, int n, double cond)
     int fails = 0;
     if (info != 0) {
         ++fails;
-        std::printf("    info = %d (expected 0)\n", (int)info);
+        std::printf("    info = %ld (expected 0)\n", (long)info);
     }
 
     double worst_res = 0, worst_el = 0, worst_untouched = 0;
@@ -223,7 +223,7 @@ template <class T> int suite3(int nm, int n, int nrhs)
     int fails = 0;
     if (info != 0) {
         ++fails;
-        std::printf("    info = %d (expected 0)\n", (int)info);
+        std::printf("    info = %ld (expected 0)\n", (long)info);
     }
     double worst_fwd = 0, worst_res = 0;
     std::vector<T> AX(sB);
